requirements.txt
Flask>=2.0
flask-cors
python-dotenv
requests
firebase-admin
waitress
